requirements.txt
streamlit
opencv-python
numpy
pillow
pandas
plotly
mysql-connector-python
